Fix file paths ToSourceName()
Some models like this Japanese wooden lamp ( https://quixel.com/assets/vevjfb3qx ) generate paths with spaces and Uppercase characters, things the Source Engine doesn't like.

At the time of export from Quixel the process works as intended, but the things break if you open the new assets in modeldoc or in the material editor.
this should fix that

also some paths in the final assets contained some // in the paths, the things seems to work fine with that, but it bothers me to see that in modeldoc and the material editor so fixed that too.

Aaannnd.... also deleted a couple of trailing spaces in some files.